split experience up more
