+++ apps/playwright/tests/applicant/experience/community.spec.ts
@@ -2,20 +2,19 @@ import { test, expect } from "~/fixtures";
 import ExperiencePage from "~/fixtures/ExperiencePage";
 import { loginBySub } from "~/utils/auth";
 
-test.describe("Community experience", () => {
-  const uniqueTestId = Date.now().valueOf();
-  test("Can create community experience", async ({ appPage }) => {
-    const title = `Test add community experience (${uniqueTestId})`;
-    const experiencePage = new ExperiencePage(appPage.page);
-    await loginBySub(experiencePage.page, "applicant@test.com");
+const uniqueTestId = Date.now().valueOf();
 
-    await experiencePage.addCommunityExperience({
-      title,
-      startDate: "2001-01",
-    });
+test("Can create community experience", async ({ appPage }) => {
+  const title = `Test add community experience (${uniqueTestId})`;
+  const experiencePage = new ExperiencePage(appPage.page);
+  await loginBySub(experiencePage.page, "applicant@test.com");
 
-    await expect(experiencePage.page.getByRole("alert")).toContainText(
-      /successfully added experience/i,
-    );
+  await experiencePage.addCommunityExperience({
+    title,
+    startDate: "2001-01",
   });
+
+  await expect(experiencePage.page.getByRole("alert")).toContainText(
+    /successfully added experience/i,
+  );
 });
