+++ apps/playwright/tests/applicant/experience/education.spec.ts
@@ -2,17 +2,15 @@ import { test, expect } from "~/fixtures";
 import ExperiencePage from "~/fixtures/ExperiencePage";
 import { loginBySub } from "~/utils/auth";
 
-test.describe("Education experience", () => {
-  test("Can create education experience", async ({ appPage }) => {
-    const experiencePage = new ExperiencePage(appPage.page);
-    await loginBySub(experiencePage.page, "applicant@test.com");
+test("Can create education experience", async ({ appPage }) => {
+  const experiencePage = new ExperiencePage(appPage.page);
+  await loginBySub(experiencePage.page, "applicant@test.com");
 
-    await experiencePage.addEducationExperience({
-      startDate: "2001-01",
-    });
-
-    await expect(experiencePage.page.getByRole("alert")).toContainText(
-      /successfully added experience/i,
-    );
+  await experiencePage.addEducationExperience({
+    startDate: "2001-01",
   });
+
+  await expect(experiencePage.page.getByRole("alert")).toContainText(
+    /successfully added experience/i,
+  );
 });
