+++ apps/playwright/tests/applicant/experience/work-external.spec.ts
@@ -2,21 +2,19 @@ import { test, expect } from "~/fixtures";
 import ExperiencePage from "~/fixtures/ExperiencePage";
 import { loginBySub } from "~/utils/auth";
 
-test.describe("External work experience", () => {
-  const uniqueTestId = Date.now().valueOf();
+const uniqueTestId = Date.now().valueOf();
 
-  test("Can create external work experience", async ({ appPage }) => {
-    const role = `Test add external work experience (${uniqueTestId})`;
-    const experiencePage = new ExperiencePage(appPage.page);
-    await loginBySub(experiencePage.page, "applicant@test.com");
+test("Can create external work experience", async ({ appPage }) => {
+  const role = `Test add external work experience (${uniqueTestId})`;
+  const experiencePage = new ExperiencePage(appPage.page);
+  await loginBySub(experiencePage.page, "applicant@test.com");
 
-    await experiencePage.addExternalWorkExperience({
-      role,
-      startDate: "2001-01",
-    });
-
-    await expect(experiencePage.page.getByRole("alert")).toContainText(
-      /successfully added experience/i,
-    );
+  await experiencePage.addExternalWorkExperience({
+    role,
+    startDate: "2001-01",
   });
+
+  await expect(experiencePage.page.getByRole("alert")).toContainText(
+    /successfully added experience/i,
+  );
 });
