+++ apps/playwright/tests/applicant/experience/work-student.spec.ts
@@ -2,18 +2,20 @@ import { test, expect } from "~/fixtures";
 import ExperiencePage from "~/fixtures/ExperiencePage";
 import { loginBySub } from "~/utils/auth";
 
-test("Can link skill to experience", async ({ appPage }) => {
+const uniqueTestId = Date.now().valueOf();
+
+test("Can create goc student work experience", async ({ appPage }) => {
+  const role = `Test add goc student work experience (${uniqueTestId})`;
   const experiencePage = new ExperiencePage(appPage.page);
   await loginBySub(experiencePage.page, "applicant@test.com");
 
-  const skill = "Courage";
-
-  await experiencePage.linkSkillToExperience({
-    experienceType: "work",
-    skill: skill,
+  await experiencePage.addGovStudentWorkExperience({
+    role,
+    startDate: "2001-01",
+    endDate: "2200-01",
   });
 
   await expect(experiencePage.page.getByRole("alert")).toContainText(
-    skill + " selected.",
+    /successfully added experience/i,
   );
 });
