+++ apps/playwright/tests/applicant/experience/work-term.spec.ts
@@ -4,14 +4,17 @@ import { loginBySub } from "~/utils/auth";
 
 const uniqueTestId = Date.now().valueOf();
 
-test("Can create award experience", async ({ appPage }) => {
-  const title = `Test add award experience (${uniqueTestId})`;
+test("Can create goc term or indeterminate work experience", async ({
+  appPage,
+}) => {
+  const role = `Test add goc term or indeterminate work experience (${uniqueTestId})`;
   const experiencePage = new ExperiencePage(appPage.page);
   await loginBySub(experiencePage.page, "applicant@test.com");
 
-  await experiencePage.addAwardExperience({
-    title,
-    awardedDate: "2001-01",
+  await experiencePage.addGovTermOrIndeterminateWorkExperience({
+    role,
+    startDate: "2001-01",
+    endDate: "2023-01",
   });
 
   await expect(experiencePage.page.getByRole("alert")).toContainText(
